feat : SignServiceImpl sign 예외처리

+++ src/main/java/org/knulikelion/challengers_backend/service/Impl/SignServiceImpl.java
@@ -12,6 +12,7 @@ import org.knulikelion.challengers_backend.data.entity.EmailVerification;
 import org.knulikelion.challengers_backend.data.entity.User;
 import org.knulikelion.challengers_backend.data.repository.EmailVerificationRepository;
 import org.knulikelion.challengers_backend.data.repository.UserRepository;
+import org.knulikelion.challengers_backend.service.Exception.UserNotFoundException;
 import org.knulikelion.challengers_backend.service.MailService;
 import org.knulikelion.challengers_backend.service.SignService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -22,7 +23,6 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
 
 @Slf4j
 @Service
@@ -49,10 +49,14 @@ public class SignServiceImpl implements SignService {
         User findUser = userRepository.findByEmail(email+domain);
         BaseResponseDto baseResponseDto = new BaseResponseDto();
 
-        if(findUser != null){
+        if(findUser != null && findUser.isUseAble()){
             baseResponseDto.setSuccess(false);
             baseResponseDto.setMsg("이미 존재하는 계정입니다.");
-        }else{
+
+        } else if (findUser !=null && !findUser.isUseAble()) {
+            baseResponseDto.setSuccess(false);
+            baseResponseDto.setMsg("챌린져스 서비스에 가입했던 계정이 있습니다. 관리자에게 문의하세요.");
+        }else {
             baseResponseDto.setSuccess(true);
             baseResponseDto.setMsg("회원 가입이 가능합니다.");
         }
@@ -62,30 +66,46 @@ public class SignServiceImpl implements SignService {
     public BaseResponseDto sendCode(SignUpRequestDto signUpRequestDto) {
         BaseResponseDto baseResponseDto = new BaseResponseDto();
 
-        List<EmailVerification> checkEmail = emailVerificationRepository.findByEmail(signUpRequestDto.getEmail()+domain);
-        if(!checkEmail.isEmpty()){
-            for(EmailVerification temp : checkEmail) {
-                emailVerificationRepository.delete(temp);
+        User user = userRepository.findByEmail(signUpRequestDto.getEmail() + domain);
+
+        if (user != null && user.isUseAble()) { /*신규 유저가 아닐 때*/
+            log.info("[sendCode] : {}",user.isUseAble());
+            return BaseResponseDto.builder()
+                    .success(false)
+                    .msg("챌린져스에 가입한 계정이 존재합니다!")
+                    .build();
+
+        } else if (user != null && !user.isUseAble()) {
+            return BaseResponseDto.builder()
+                    .success(false)
+                    .msg("챌린져스에 가입했던 계정이 있습니다. 관리자에게 문의하세요.")
+                    .build();
+        } else {
+            List<EmailVerification> checkEmail = emailVerificationRepository.findByEmail(signUpRequestDto.getEmail() + domain);
+            if (!checkEmail.isEmpty()) {
+                for (EmailVerification temp : checkEmail) {
+                    emailVerificationRepository.delete(temp);
+                }
             }
+
+            log.info("[sendCode] 인증 번호 전송 Email : {}", signUpRequestDto.getEmail() + domain);
+            String approvalNumber = mailService.sendMail(signUpRequestDto.getEmail() + domain);
+            log.info("[sendCode] 인증 번호 전송 완료 Email : {}, number : {}", signUpRequestDto.getEmail(), approvalNumber);
+
+            EmailVerification emailVerification = EmailVerification.builder()
+                    .email(signUpRequestDto.getEmail() + domain)
+                    .code(approvalNumber)
+                    .expireTime(LocalDateTime.now().plusMinutes(5)) /*현재 시간으로부터 5분 뒤*/
+                    .build();
+
+            emailVerificationRepository.save(emailVerification);
+            log.info("[sendCode] DataBase 저장 완료");
+
+            baseResponseDto.setSuccess(true);
+            baseResponseDto.setMsg("성공적으로 인증 번호를 전송하였습니다.");
+
+            return baseResponseDto;
         }
-
-        log.info("[sendCode] 인증 번호 전송 Email : {}", signUpRequestDto.getEmail()+domain);
-        String approvalNumber = mailService.sendMail(signUpRequestDto.getEmail() + domain);
-        log.info("[sendCode] 인증 번호 전송 완료 Email : {}, number : {}", signUpRequestDto.getEmail(), approvalNumber);
-
-        EmailVerification emailVerification = EmailVerification.builder()
-                .email(signUpRequestDto.getEmail()+domain)
-                .code(approvalNumber)
-                .expireTime(LocalDateTime.now().plusMinutes(5)) /*현재 시간으로부터 5분 뒤*/
-                .build();
-
-        emailVerificationRepository.save(emailVerification);
-        log.info("[sendCode] DataBase 저장 완료");
-
-        baseResponseDto.setSuccess(true);
-        baseResponseDto.setMsg("성공적으로 인증 번호를 전송하였습니다.");
-
-        return baseResponseDto;
     }
 
 
@@ -183,25 +203,31 @@ public class SignServiceImpl implements SignService {
     public SignInResponseDto signIn(SignInRequestDto signInRequestDto) {
         log.info("[getSignInResult] signDataHandler 로 회원 정보 요청");
         User user = userRepository.getByEmail(signInRequestDto.getEmail());
-        log.info("[getSignInResult] Id : {}", signInRequestDto.getEmail());
 
-        log.info("[getSignInResult] 패스워드 비교 수행");
-        if (!passwordEncoder.matches(signInRequestDto.getPassword(), user.getPassword())) {
-            throw new RuntimeException();
+        if(user.isUseAble()) {
+
+            log.info("[getSignInResult] Id : {}", signInRequestDto.getEmail());
+
+            log.info("[getSignInResult] 패스워드 비교 수행");
+            if (!passwordEncoder.matches(signInRequestDto.getPassword(), user.getPassword())) {
+                throw new RuntimeException();
+            }
+            log.info("[getSignInResult] 패스워드 일치");
+            log.info("[getSignInResult] SignInResultDto 객체 생성");
+            SignInResponseDto signInResultDto = SignInResponseDto.builder()
+                    .accessToken(jwtTokenProvider.createAccessToken(String.valueOf(user.getEmail()), user.getRoles()))
+                    .refreshToken(jwtTokenProvider.createRefreshToken(String.valueOf(user.getEmail())))
+                    .email(user.getEmail())
+                    .userName(user.getUserName())
+                    .code(1)
+                    .msg("로그인에 성공하였습니다.")
+                    .build();
+            log.info("[getSignInResult] SignInResultDto 객체에 값 주입");
+
+            return signInResultDto;
+        }else{
+            throw new UserNotFoundException("사용할 수 없는 사용자 계정입니다. 관리자에게 문의하세요.");
         }
-        log.info("[getSignInResult] 패스워드 일치");
-        log.info("[getSignInResult] SignInResultDto 객체 생성");
-        SignInResponseDto signInResultDto = SignInResponseDto.builder()
-                .accessToken(jwtTokenProvider.createAccessToken(String.valueOf(user.getEmail()), user.getRoles()))
-                .refreshToken(jwtTokenProvider.createRefreshToken(String.valueOf(user.getEmail())))
-                .email(user.getEmail())
-                .userName(user.getUserName())
-                .code(1)
-                .msg("로그인에 성공하였습니다.")
-                .build();
-        log.info("[getSignInResult] SignInResultDto 객체에 값 주입");
-
-        return signInResultDto;
     }
 
 }
